Support User and Org for a valid session
Currently we're creating an unexpected user and organization when
creating a session. This is acceptable when you don't need access to
either, but in many situations it's beneficial to be able to explicitly
setup the organization and user for your test environment.

I've added support for passing in both of these values, while
continuing to support the current default by using mocks.